armv8: layerscape: don't remove crypto node on LS1028A
At least on the LS1028A the crypto module is always there, but has some
functions disabled on non-E parts. Thus it doesn't make sense to
completely remove the node. Linux will figure out what is there and what
is not.

Please note, this patch is certainly not ideal. But I don't know which
processors have the SEC module always available. NXP would have to
comment on that and make sure, the condition when to remove the crypto
node is correct.

Signed-off-by: Michael Walle <michael@walle.cc>